Added general library for generic methods
