--- src/config.js
@@ -1,17 +1,0 @@
-import * as Phaser from 'phaser';
-
-const CONFIG = {
-  type: Phaser.AUTO,
-  width: 800,
-  height: 600,
-  parent: 'game-container',
-  physics: {
-    default: 'arcade',
-    arcade: {
-      debug: false,
-      gravity: { y: 0 },
-    },
-  },
-};
-
-export default CONFIG;
