+++ src/index.html
@@ -8,6 +8,7 @@
   <title>SPACE⭐️CRAFT</title>
 
   <script src="app.bundle.js" defer></script>
+  <link rel="icon" type="image/png" href="assets/sprites/pilot.png">
   <link rel="stylesheet" href="assets/styles/site.css">
 </head>
