+++ src/keys/api.js
@@ -1,5 +1,5 @@
 const API_KEYS = {
-  LEADERBOARD: 'LEADERBOARD',
+  LEADERBOARD: 'isFb6OL6cyCDWu7w3e7i',
 };
 
 export default API_KEYS;
