--- src/keys/assets.js
@@ -1,23 +1,0 @@
-const ASSETS_KEYS = {
-  BATTLE_CRUISER: 'battlecruiser',
-  BEAM: 'beam',
-  GAME_OVER: 'gameover',
-  TOP_TEN: 'top10',
-  CREDITS: 'credits',
-  INPUT_NAME: 'inputName',
-  LIFE: 'life',
-  PILOT: 'pilot',
-  SPACE_BACKGROUND: 'spaceBackground',
-  LOGO: 'phaserLogo',
-  BUTTON: 'button',
-  BUTTON_ON_HOVER: 'buttonOnHover',
-  UNCHECKED_BOX: 'uncheckedBox',
-  CHECKED_BOX: 'checkedBox',
-  ASTEROID: 'asteroid',
-  EXPLOSION: 'explosion',
-  BG_MUSIC: 'bgMusic',
-  EXPLOSION_SOUND: 'exposionAudio',
-  PIXEL_FONT: 'pixelFont',
-};
-
-export default ASSETS_KEYS;
