--- src/keys/global.js
@@ -1,5 +1,0 @@
-const GLOBAL_VAR = {
-  MUSIC_ON: false,
-};
-
-export default GLOBAL_VAR;
