--- src/keys/scene.js
@@ -1,10 +1,0 @@
-const SCENE_KEYS = {
-  GAME_START: 'Game_Start',
-  GAME_OVER: 'Game_Over',
-  CREDITS: 'Credits',
-  PRELOADER: 'Preloader',
-  TITLE: 'Title',
-  TOP10SCORES: 'Top10Scores',
-};
-
-export default SCENE_KEYS;
